+++ build.zig
@@ -351,6 +351,7 @@ pub fn build(b: *std.Build) void {
         "tests/cli/local_import_test.zig",
         "tests/cli/lsp_test.zig",
         "tests/cli/sqlite_test.zig",
+        "tests/cli/shared_mem_test.zig",
         "tests/cli/help_screen_test.zig",
         "tests/fuzz/lexer_parser_checker_fuzz.zig",
         "tests/golden/golden_test.zig",